TINY-10324: getMatches
