+++ modules/tinymce/src/core/main/ts/textpatterns/core/BlockPattern.ts
@@ -1,10 +1,13 @@
-import { Arr } from '@ephox/katamari';
+import { Arr, Optional } from '@ephox/katamari';
 import { SugarText, SugarElement } from '@ephox/sugar';
 
 import DOMUtils from '../../api/dom/DOMUtils';
 import Editor from '../../api/Editor';
+import * as InlinePattern from '../core/InlinePattern';
+import { PatternSet } from '../core/PatternTypes';
+import * as Utils from '../utils/Utils';
 import * as BlockPatternUtils from './BlockPatternUtils';
-import { BlockPattern, BlockPatternMatch } from './PatternTypes';
+import { BlockPattern, BlockPatternMatch, InlinePatternMatch } from './PatternTypes';
 
 const startsWithSingleSpace = (s: string): boolean => /^\s[^\s]/.test(s);
 
@@ -32,6 +35,19 @@ const findPattern = BlockPatternUtils.findPattern((pattern, text, nuText) => tex
 
 const findPatterns = BlockPatternUtils.createFindPatterns(findPattern, true);
 
+const getMatches = (editor: Editor, patternSet: PatternSet): Optional<{ inlineMatches: InlinePatternMatch[]; blockMatches: BlockPatternMatch[] }> => {
+  const rng = editor.selection.getRng();
+  return Utils.getParentBlock(editor, rng).map((block) => {
+    const offset = Math.max(0, rng.startOffset);
+    const dynamicPatternSet = Utils.resolveFromDynamicPatterns(patternSet, block, block.textContent ?? '');
+    // IMPORTANT: We need to get normalized match results since undoing and redoing the editor state
+    // via undoManager.extra() will result in the DOM being normalized.
+    const inlineMatches = InlinePattern.findPatterns(editor, block, rng.startContainer, offset, dynamicPatternSet, true);
+    const blockMatches = findPatterns(editor, block, dynamicPatternSet, true);
+    return { inlineMatches, blockMatches };
+  }).filter(({ inlineMatches, blockMatches }) => blockMatches.length > 0 || inlineMatches.length > 0);
+};
+
 const applyMatches = (editor: Editor, matches: BlockPatternMatch[]): void => {
   if (matches.length === 0) {
     return;
@@ -43,4 +59,4 @@ const applyMatches = (editor: Editor, matches: BlockPatternMatch[]): void => {
   editor.selection.moveToBookmark(bookmark);
 };
 
-export { applyMatches, findPattern, findPatterns };
+export { applyMatches, findPattern, findPatterns, getMatches };
