+++ modules/tinymce/src/core/main/ts/textpatterns/keyboard/KeyHandler.ts
@@ -1,4 +1,4 @@
-import { Unicode } from '@ephox/katamari';
+import { Fun, Unicode } from '@ephox/katamari';
 
 import * as TextSearch from '../../alien/TextSearch';
 import Editor from '../../api/Editor';
@@ -10,16 +10,10 @@ import * as InlinePattern from '../core/InlinePattern';
 import { PatternSet } from '../core/PatternTypes';
 import * as Utils from '../utils/Utils';
 
-const handleEnter = (editor: Editor, patternSet: PatternSet): boolean => {
-  const rng = editor.selection.getRng();
-  return Utils.getParentBlock(editor, rng).map((block) => {
-    const offset = Math.max(0, rng.startOffset);
-    const dynamicPatternSet = Utils.resolveFromDynamicPatterns(patternSet, block, block.textContent ?? '');
-    // IMPORTANT: We need to get normalized match results since undoing and redoing the editor state
-    // via undoManager.extra() will result in the DOM being normalized.
-    const inlineMatches = InlinePattern.findPatterns(editor, block, rng.startContainer, offset, dynamicPatternSet, true);
-    const blockMatches = BlockPattern.findPatterns(editor, block, dynamicPatternSet, true);
-    if (blockMatches.length > 0 || inlineMatches.length > 0) {
+const handleEnter = (editor: Editor, patternSet: PatternSet): boolean =>
+  BlockPattern.getMatches(editor, patternSet).fold(
+    Fun.never,
+    ({ inlineMatches, blockMatches }) => {
       editor.undoManager.add();
       editor.undoManager.extra(
         () => {
@@ -45,10 +39,7 @@ const handleEnter = (editor: Editor, patternSet: PatternSet): boolean => {
         }
       );
       return true;
-    }
-    return false;
-  }).getOr(false);
-};
+    });
 
 const handleInlineKey = (
   editor: Editor,
@@ -68,22 +59,15 @@ const handleInlineKey = (
   });
 };
 
-const handleBlockPatternOnSpace = (editor: Editor, patternSet: PatternSet): boolean => {
-  const rng = editor.selection.getRng();
-  return Utils.getParentBlock(editor, rng).exists((block) => {
-    const offset = Math.max(0, rng.startOffset);
-    const beforeText = Utils.getBeforeText(editor.dom, block, rng.startContainer, offset);
-    const dynamicPatternSet = Utils.resolveFromDynamicPatterns(patternSet, block, beforeText);
-    const matches = BlockPatternOnSpace.findPatterns(editor, block, dynamicPatternSet, false, beforeText);
-    const hasMatches = matches.length > 0;
-    if (hasMatches) {
+const handleBlockPatternOnSpace = (editor: Editor, patternSet: PatternSet): boolean =>
+  BlockPatternOnSpace.getMatches(editor, patternSet).fold(
+    Fun.never,
+    (matches) => {
       editor.undoManager.transact(() => {
         BlockPatternOnSpace.applyMatches(editor, matches);
       });
-    }
-    return hasMatches;
-  });
-};
+      return true;
+    });
 
 const checkKeyEvent = <T>(codes: T[], event: KeyboardEvent, predicate: (code: T, event: KeyboardEvent) => boolean): boolean => {
   for (let i = 0; i < codes.length; i++) {
